luna_v.3.3.3/luna-api/requirements.txt
sqlalchemy-utils==0.32.13
SQLAlchemy-migrate==0.11.0
Requests==2.13.0
pillow-simd==4.3.0.post0
psycopg2==2.7.1
enum34==1.1.6
logbook==1.1.0
ujson==1.35
tornado==5.0.1
SQLAlchemy==1.2.5
faker==0.7.10
influxdb==4.1.0
pluginbase==0.5
passlib==1.7.1
six==1.11.0
python-dateutil==2.7.3
cx_oracle==6.3.1
pytz==2018.4
jsonschema==2.6.0
chardet==3.0.4
alembic==1.0.0
extras/luna3-v.0.0.54.tar.gz
extras/LunaVL-v.1.0.11.tar.gz
